🔒🚸 improve authentication (#137)
* 🔒 remove auth local storage cache

Caching tokens in local storage is not recommended. See https://auth0.com/docs/libraries/auth0-single-page-app-sdk#change-storage-options. I've replaced the cache by a silent authentication. The app tries a silent login first, and then if it is denied by Auth0, it falls back to the normal interactive login sequence.

* 🚸 go straight to google for login

Skip the Auth0 login prompt and go straight to Google when logging in, since this is the only connection available anyway.